+++ .github/workflows/integ.yml
@@ -5,7 +5,7 @@ on:
   pull_request: {}
   workflow_dispatch: {}
 jobs:
-  integ-init:
+  init:
     runs-on: ubuntu-latest
     permissions:
       contents: read
@@ -14,7 +14,7 @@ jobs:
       - name: Set up Node.js
         uses: actions/setup-node@v2
         with:
-          node-version: 14
+          node-version: "14"
       - name: Set up Python 3.x
         uses: actions/setup-python@v2
         with:
@@ -29,7 +29,7 @@ jobs:
         run: yarn install --frozen-lockfile
       - name: Run integration tests
         run: yarn run integ:init
-  integ-init-typescript-app:
+  init-typescript-app:
     runs-on: ubuntu-latest
     permissions:
       contents: read
